It creates a data object for Non Parametric Quality Control

It creates an object of class 'npqcd' to perform statistical quality control. This object is used to plot Non Parametric Multivariate Control Charts.

npqcd(x, G = NULL, data.name = NULL)

Arguments

x
a matrix or data-frame or array which it should contain data. Dimension has to be the same as that of the observations.

G
The x as a matrix, data frame or list. If it is a matrix or data frame, then each row is viewed as one multivariate observation.

data.name
a string that specifies the title displayed on the plots. If not provided it is taken from the name of the object x.

Examples

library(qcr)

set.seed(356)
data <- matrix(rnorm(999), nc = 3)
x <-rexp(999,0.5) 
x <-matrix(x,ncol=3) 
data.npqcd <- npqcd(data,x)
str(data.npqcd)

#> List of 2
#>  $ x: num [1:333, 1:3, 1] 0.306 -1.179 -0.322 -0.258 1.389 ...
#>  $ G: num [1:333, 1:3] 1.75 1.47 5.4 0.97 1.55 ...
#>  - attr(*, "data.name")= chr "DATA"
#>  - attr(*, "type.data")= chr "Multivariate"
#>  - attr(*, "class")= chr [1:2] "npqcd" "list"