Function to create a plotting 'qcs' object

Generic function for plotting Shewhart charts of object of class 'qcs' to perform statistical quality control.

Usage

# S3 method for qcs
plot(
  x,
  title,
  subtitle,
  xlab,
  ylab,
  ylim,
  center.nominal = NULL,
  limits.specification = NULL,
  limits.alert = NULL,
  type.data = c("continuous", "atributte", "dependence"),
  ...
)

# S3 method for qcs.xbar
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.S
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.R
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.one
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.p
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.np
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.c
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.u
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

# S3 method for qcs.ewma
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

# S3 method for qcs.cusum
plot(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

Arguments

x

Object qcs (Quality Control Statical).

title

An overall title for the plot.

subtitle

A sub title for the plot.

xlab

A title for the x axis.

ylab

A title for the y axis.

ylim

The y limits of the plot.

center.nominal

A value specifying the center of group statistics or the "target" value of the process.

limits.specification

A two-value vector specifying control limits.

limits.alert

A two-value vector specifying control alert limits.

type.data

A string specifying the type of data.

...

Arguments to be passed to or from methods.

conf.nsigma.alert

A numeric value used to compute control limits, specifying the number of standard deviations (if conf.nsigma > 1) or the confidence level (if 0 < conf.nsigma < 1).

conf.nsigma

A numeric value used to compute control limits, specifying the number of standard deviations (if conf.nsigma > 1) or the confidence level (if 0 < conf.nsigma < 1).