Quality Control Data

Create an object of class 'qcd' to perform statistical quality control. This object may then be used to plot Shewhart charts, Multivariate Control Charts, and more.

Usage

qcd(
  data,
  var.index = 1,
  sample.index = 2,
  covar.index = NULL,
  covar.names = NULL,
  data.name = NULL,
  type.data = c("continuous", "atributte", "dependence"),
  sizes = NULL
)

Arguments

data

A matrix or data-frame which should contain data, index sample and, optionally, covariate(s).

var.index

A scalar with the column number corresponding to the observed data for the variable (the variable quality). Alternativelly can be a string with the name of the quality variable.

sample.index

A scalar with the column number corresponding to the index fo each group (sample).

covar.index

Optional. A scalar or numeric vector with the column number(s) corresponding to the covariate(s). Alternativelly it can be a character vector with the names of the covariates.

covar.names

Optional. A string or vector of strings indicating the names for the covariate columns. Only valid if there is more than one column of data. By default, takes the names from the original object.

data.name

A string specifying the name of the variable which appears on the plots. If not provided, it is taken from the object given as data.

type.data

A string specifying the type of data.

sizes

Optional. A value or a vector of values specifying the sample sizes associated with each group. For continuous data, the sample sizes are obtained counting the non-NA elements of the sample.index vector. For attribute variable the argument sizes is required.

Examples

if (FALSE) {
library(qcr)
data(pistonrings)
str(pistonrings)
pistonrings.qcd<-qcd(pistonrings)
class(pistonrings.qcd)
}