Quality Control Statistics

Create an object of class 'qcs' to perform statistical quality control. This object may then be used to plot Shewhart charts, Multivariate Control Charts, and more.

qcs(
  x,
  sample.index,
  sizes = NULL,
  type = c("xbar", "R", "S", "one", "p", "np", "c", "u", "ewma", "cusum"),
  center = NULL,
  std.dev,
  conf.nsigma = 3,
  limits = NULL,
  type.data = c("continuous", "atributte", "dependence"),
  lambda = 0.2,
  decision.interval = 5,
  se.shift = 1
)

qcs.continuous(
  x,
  sample.index,
  sizes = NULL,
  type = c("xbar", "R", "S", "one"),
  center = NULL,
  std.dev,
  conf.nsigma = 3,
  limits = NULL
)

qcs.atributte(
  x,
  sample.index = NULL,
  sizes = NULL,
  type = c("p", "np", "c", "u"),
  center = NULL,
  conf.nsigma = 3,
  limits = NULL
)

qcs.dependence(
  x,
  sample.index = NULL,
  sizes = NULL,
  type = c("ewma", "cusum"),
  center = NULL,
  std.dev,
  nsigma = 3,
  lambda = 0.2,
  decision.interval = 5,
  se.shift = 1
)

Arguments

x
a vector containing observed data

sample.index
a scalar with the column number corresponding to the index of each group (sample).

sizes
a value or a vector of values specifying the sample sizes associated with each group. For continuous data the sample sizes are obtained counting the non-NA elements of the sample.index vector. For "p", "np" and "u" charts the argument sizes is required.

type
a character string specifying the group statistics to compute:

Statistic charted	Chart description

"xbar"	mean	means of a continuous process variable

"R"	range	ranges of a continuous process variable

"S"	standard deviation	standard deviations of a continuous variable

"one"	mean	one-at-time data of a continuous process variable

"p"	proportion	proportion of nonconforming units

"np"	count	number of nonconforming units

"c"	count	nonconformities per unit

"u"	count	average nonconformities per unit

"g"	count	number of non-events between events

center
a value specifying the center of group statistics or the ''target'' value of the process.

std.dev
a value or an available method specifying the within-group standard deviation(s) of the process. Several methods are available for estimating the standard deviation in case of a continuous process variable.

conf.nsigma
a numeric value used to compute control limits, specifying the number of standard deviations (if conf.nsigma > 1) or the confidence level (if 0 < conf.nsigma < 1).

limits
a two-value vector specifying control limits.

type.data
a string specifying el type de data.

lambda
the smoothing parameter \(0 \le \lambda \le 1\)

decision.interval
A numeric value specifying the number of standard errors of the summary statistics at which the cumulative sum is out of control.

se.shift
The amount of shift to detect in the process, measured in standard errors of the summary statistics.

nsigma
a numeric value used to compute control limits, specifying the number of standard deviations.

Value

Returns an object of class 'qcs'.

References

Montgomery, D.C. (2000) Introduction to Statistical Quality Control, 4th ed. New York: John Wiley & Sons.
Wetherill, G.B. and Brown, D.W. (1991) Statistical Process Control. New York: Chapman & Hall.